opentelemetry/python-app/requirements.txt
fastapi<1
flask<3
gunicorn<21
honcho<2
meysam-utils[settings]<1
opentelemetry-distro==0.37b0
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-asgi==0.37b0
opentelemetry-instrumentation-celery==0.37b0
opentelemetry-instrumentation-fastapi==0.37b0
opentelemetry-instrumentation-logging==0.37b0
opentelemetry-instrumentation-pymongo==0.37b0
opentelemetry-instrumentation-redis==0.37b0
opentelemetry-instrumentation-system-metrics==0.37b0
uvicorn<1
